requirements.txt
flask==2.3.2
requests==2.31.0
pandas==2.0.2
python-dotenv==1.0.0
google-api-python-client==2.89.0
google-auth-oauthlib==1.0.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
gslides==0.1.1
